Generate a ranking for each month
This is a bit of mess and needs to be cleaned up
later, but for now it:
- Adds 3 new fields to the data table
- Generates rankings for each month in a period
- Bases ordering of graphs/tables on month3

We should be able to remove period_average and
ranking_average from database soon.

The method for producing the ranking is crap and
will probably perform really poorly. We should
fix at some point.